+++ src/features/XIT/HELP.ts
@@ -1,8 +1,0 @@
-import HELP from '@src/features/XIT/HELP.vue';
-
-xit.add({
-  command: 'HELP',
-  name: 'HELP',
-  description: 'Useful information to get started with Refined PrUn.',
-  component: () => HELP,
-});