+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- `INV`: Fixed a "Component failed to render" error on ship flight start (`inv-ship-destination-location`)
 - `XIT ACT`: Fixed an error when some ships are not named
 
 ## 26.9.15